refactor add web app
majore updates
