--- multi-tenant-jwt-resourceserver/complex/src/main/java/ca/simplestep/example/resourceserver/jwt/complex/config/ResourceServerConfig.java
@@ -1,56 +1,0 @@
-package ca.simplestep.example.resourceserver.jwt.complex.config;
-
-import ca.simplestep.example.resourceserver.jwt.complex.config.tenant.Tenant;
-import ca.simplestep.example.resourceserver.jwt.complex.config.tenant.TenantRepository;
-import com.nimbusds.jose.proc.SecurityContext;
-import com.nimbusds.jwt.proc.ConfigurableJWTProcessor;
-import com.nimbusds.jwt.proc.DefaultJWTProcessor;
-import com.nimbusds.jwt.proc.JWTClaimsSetAwareJWSKeySelector;
-import com.nimbusds.jwt.proc.JWTProcessor;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.oauth2.core.DelegatingOAuth2TokenValidator;
-import org.springframework.security.oauth2.core.OAuth2TokenValidator;
-import org.springframework.security.oauth2.jwt.Jwt;
-import org.springframework.security.oauth2.jwt.JwtDecoder;
-import org.springframework.security.oauth2.jwt.JwtValidators;
-import org.springframework.security.oauth2.jwt.NimbusJwtDecoder;
-
-@Configuration
-@EnableConfigurationProperties(ApplicationProperties.class)
-public class ResourceServerConfig {
-
-    private ApplicationProperties properties;
-
-    public ResourceServerConfig(ApplicationProperties properties) {
-        this.properties = properties;
-    }
-
-    @Bean
-    TenantRepository inMemoryTenantRepository() {
-        TenantRepository tenantRepository = new TenantRepository();
-        properties.getIssuers().forEach(i -> tenantRepository.save(new Tenant(i)));
-        return tenantRepository;
-    }
-
-    @Bean
-    JWTProcessor<SecurityContext> jwtProcessor(JWTClaimsSetAwareJWSKeySelector<SecurityContext> keySelector) {
-        ConfigurableJWTProcessor<SecurityContext> jwtProcessor =
-                new DefaultJWTProcessor<SecurityContext>();
-        jwtProcessor.setJWTClaimsSetAwareJWSKeySelector(keySelector);
-        return jwtProcessor;
-    }
-
-    @Bean
-    JwtDecoder jwtDecoder(JWTProcessor<SecurityContext> jwtProcessor, OAuth2TokenValidator<Jwt> jwtValidator) {
-        NimbusJwtDecoder decoder = new NimbusJwtDecoder(jwtProcessor);
-        OAuth2TokenValidator<Jwt> validator = new DelegatingOAuth2TokenValidator<>(
-                JwtValidators.createDefault(),
-                jwtValidator
-                );
-        decoder.setJwtValidator(validator);
-        return decoder;
-    }
-}
