--- multi-tenant-jwt-resourceserver/complex/src/main/java/ca/simplestep/example/resourceserver/jwt/complex/config/tenant/Tenant.java
@@ -1,18 +1,0 @@
-package ca.simplestep.example.resourceserver.jwt.complex.config.tenant;
-
-import ca.simplestep.example.resourceserver.jwt.complex.config.tenant.utils.OidcProviderConfigurationUtils;
-
-import java.util.Map;
-
-public class Tenant {
-
-    private final Map<String, Object> attributes;
-
-    public Tenant(String issuerUri) {
-        attributes = OidcProviderConfigurationUtils.getConfigurationForOidcIssuerLocation(issuerUri);
-    }
-
-    public Object getAttribute(String attribute) {
-        return attributes.get(attribute);
-    }
-}
