+++ multi-tenant-jwt-resourceserver/complex/src/main/java/com/cedarlogic/jwt/complex/tenant/TenantJwtIssuerValidator.java
@@ -1,0 +1,38 @@
+package com.cedarlogic.jwt.complex.tenant;
+
+import org.springframework.security.oauth2.core.OAuth2TokenValidator;
+import org.springframework.security.oauth2.core.OAuth2TokenValidatorResult;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.security.oauth2.jwt.JwtIssuerValidator;
+import org.springframework.stereotype.Component;
+
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.ConcurrentHashMap;
+
+@Component
+public class TenantJwtIssuerValidator implements OAuth2TokenValidator<Jwt> {
+    private final TenantRepository tenantRepository;
+    private final Map<String, JwtIssuerValidator> validators = new ConcurrentHashMap<>();
+
+    public TenantJwtIssuerValidator(TenantRepository tenantRepository) {
+        this.tenantRepository = tenantRepository;
+    }
+
+    @Override
+    public OAuth2TokenValidatorResult validate(Jwt token) {
+        return this.validators.computeIfAbsent(toTenant(token), this::fromTenant)
+                .validate(token);
+    }
+
+    private String toTenant(Jwt jwt) {
+        return jwt.getIssuer().toString();
+    }
+
+    private JwtIssuerValidator fromTenant(String tenant) {
+        return Optional.ofNullable(this.tenantRepository.findById(tenant))
+                .map(t -> (String)t.getAttribute("issuer"))
+                .map(JwtIssuerValidator::new)
+                .orElseThrow(() -> new IllegalArgumentException("unknown tenant"));
+    }
+}
