+++ multi-tenant-jwt-resourceserver/complex/src/main/java/com/cedarlogic/jwt/complex/tenant/TenantRepository.java
@@ -1,0 +1,17 @@
+package com.cedarlogic.jwt.complex.tenant;
+
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+
+public class TenantRepository {
+
+    private final Map<String, Tenant> tenants = new ConcurrentHashMap<>();
+
+    public Tenant findById(String tenant) {
+        return tenants.get(tenant);
+    }
+
+    public void save(Tenant tenant) {
+        tenants.put((String)tenant.getAttribute("issuer"), tenant);
+    }
+}
