+++ multi-tenant-jwt-resourceserver/pom.xml
@@ -5,21 +5,22 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.4.4</version>
+		<version>2.6.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>ca.simplestep.example</groupId>
-	<artifactId>multitenant-jwt-resourceserver</artifactId>
+	<groupId>com.cedarlogic</groupId>
+	<artifactId>multitenant-jwt</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
-	<name>multitenant-jwt-resourceserver</name>
+	<name>multitenant-jwt</name>
 	<description>Example of a multitenant JWT OAuth2 Resource Server</description>
+	<properties>
+		<java.version>11</java.version>
+		<azure.version>3.14.0</azure.version>
+	</properties>
 	<packaging>pom</packaging>
 	<modules>
 		<module>simple</module>
 		<module>complex</module>
+		<module>webapp</module>
 	</modules>
-	<properties>
-		<java.version>11</java.version>
-	</properties>
-
 </project>
