+++ multi-tenant-jwt-resourceserver/simple/pom.xml
@@ -2,15 +2,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>multitenant-jwt-resourceserver</artifactId>
-        <groupId>ca.simplestep.example</groupId>
+        <artifactId>multitenant-jwt</artifactId>
+        <groupId>com.cedarlogic</groupId>
         <version>0.0.1-SNAPSHOT</version>
     </parent>
-    <modelVersion>4.0.0</modelVersion>
-
     <artifactId>simple</artifactId>
-
     <dependencies>
         <dependency>
             <groupId>org.springframework.boot</groupId>
