+++ multi-tenant-jwt-resourceserver/simple/src/main/java/com/cedarlogic/jwt/simple/SimpleLauncher.java
@@ -1,0 +1,23 @@
+package com.cedarlogic.jwt.simple;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@SpringBootApplication
+public class SimpleLauncher {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SimpleLauncher.class, args);
+	}
+
+	@RestController
+	public static class MyController {
+		@GetMapping("/")
+		public String getHello() {
+			return "hello";
+		}
+	}
+
+}
