+++ multi-tenant-jwt-resourceserver/simple/src/main/java/com/cedarlogic/jwt/simple/config/ApplicationProperties.java
@@ -1,0 +1,20 @@
+package com.cedarlogic.jwt.simple.config;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@ConfigurationProperties(prefix = "app")
+public class ApplicationProperties {
+
+    private List<String> issuers = new ArrayList<>();
+
+    public List<String> getIssuers() {
+        return issuers;
+    }
+
+    public void setIssuers(List<String> issuers) {
+        this.issuers = issuers;
+    }
+}
