--- multi-tenant-jwt-resourceserver/simple/src/test/java/ca/simplestep/example/resourceserver/jwt/simple/SimpleLauncherTests.java
@@ -1,13 +1,0 @@
-package ca.simplestep.example.resourceserver.jwt.simple;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SimpleLauncherTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
