+++ multi-tenant-jwt-resourceserver/simple/src/test/java/com/cedarlogic/jwt/simple/SimpleLauncherTests.java
@@ -1,0 +1,13 @@
+package com.cedarlogic.jwt.simple;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SimpleLauncherTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
