+++ multi-tenant-jwt-resourceserver/webapp/.gitignore
@@ -31,5 +31,3 @@ build/
 
 ### VS Code ###
 .vscode/
-.DS_Store
-multi-tenant-jwt-resourceserver/.mvn/wrapper/maven-wrapper.jar
