+++ multi-tenant-jwt-resourceserver/webapp/src/main/java/com/cedarlogic/webapp/WebappApplication.java
@@ -1,0 +1,13 @@
+package com.cedarlogic.webapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class WebappApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(WebappApplication.class, args);
+	}
+
+}
