+++ multi-tenant-jwt-resourceserver/webapp/src/test/java/com/cedarlogic/webapp/WebappApplicationTests.java
@@ -1,0 +1,13 @@
+package com.cedarlogic.webapp;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class WebappApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
